frontend: align list launch env merging
